+++ packages/liveblocks-prosemirror/.stylelintrc.cjs
@@ -1,0 +1,6 @@
+module.exports = {
+  extends: ["stylelint-config-standard"],
+  rules: {
+    "selector-class-pattern": /^collaboration-carets__[a-z-]+$/,
+  },
+};